Create conftest module for unit tests
